pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "scikit_build_core"
authors = [
    { name = "Henry Schreiner", email = "henryfs@princeton.edu" },
]
description = "Build backend for CMake based projects"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Build Tools",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Development Status :: 4 - Beta",
    "Typing :: Typed",
]

dynamic = ["version"]

dependencies = [
    "exceptiongroup; python_version<'3.11'",
    "importlib_resources >=1.3; python_version<'3.9'",
    "packaging >=20.9",
    "tomli >=1.1; python_version<'3.11'",
    "typing_extensions >=3.10.0; python_version<'3.8'",
]
# Note: for building wheels and sdists, there are also additional dependencies
# in the pyproject extra. And cmake and possibly ninja if those are not already
# present (user controllable)

[project.optional-dependencies]
pyproject = [
    "pyproject-metadata >=0.5",
    "pathspec >=0.10.1",
]
rich = [
    "rich",
]
test = [
    "build[virtualenv]",
    "cattrs >=22.2.0",
    "gitpython",
    "importlib_metadata; python_version<'3.8'",
    "pathspec >=0.10.1",
    "pybind11",
    "pyproject-metadata >=0.5",
    "pytest >=7.0",  # 7.2+ recommended for better tracebacks with ExceptionGroup
    "pytest-subprocess >=1.5",
    "setuptools",
    "wheel",
]
cov = [
    "pytest-cov[toml]",
]
dev = [
    "build",
    "cattrs >=22.2.0",
    "pytest >=7.0",
    "pytest-subprocess",
    "rich",
]
docs = [
    "furo",
    "myst_parser >=0.13",
    "sphinx >=4.0",
    "sphinx_copybutton",
]

[project.urls]
Homepage = "https://github.com/scikit-build/scikit-build-core"
Examples = "https://github.com/scikit-build/scikit-build-core/tree/main/tests/packages"

[project.entry-points."distutils.setup_keywords"]
cmake_extensions = "scikit_build_core.setuptools.extension:cmake_extensions"
cmake_source_dir = "scikit_build_core.setuptools.extension:cmake_source_dir"


[tool.hatch]
version.source = "vcs"
build.hooks.vcs.version-file = "src/scikit_build_core/_version.py"


[tool.pytest.ini_options]
minversion = "7.0"
addopts = ["-ra", "--strict-markers", "--strict-config"]
xfail_strict = true
filterwarnings = [
    "error",
    "ignore:pkg_resources is deprecated as an API:DeprecationWarning",  # Caused by wheel in tests
    "ignore:Config variable '.*' is unset, Python ABI tag may be incorrect:RuntimeWarning",
]
log_cli_level = "info"
testpaths = ["tests"]
markers = [
    "compile: Compiles code",
    "configure: Configures CMake code",
    "integration: Full package build",
    "setuptools: Tests setuptools integration",
    "virtualenv: Needs a virtualenv",
    "isolated: Needs an isolated virtualenv",
]


[tool.mypy]
files = ["src", "tests"]
mypy_path = ["$MYPY_CONFIG_FILE_DIR/src"]
python_version = "3.7"
warn_unused_configs = true
show_error_codes = true
enable_error_code = ["ignore-without-code", "redundant-expr", "truthy-bool"]
strict = true
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = ["scikit_build_core.*"]
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["pathspec", "setuptools_scm", "hatch_fancy_pypi_readme"]
ignore_missing_imports = true


[tool.pylint]
py-version = "3.7"
jobs = "0"
reports.output-format = "colorized"
similarities.ignore-imports = "yes"
good-names = ["f"]
messages_control.disable = [
    "design",
    "fixme",
    "import-outside-toplevel",
    "invalid-name",
    "line-too-long",
    "missing-class-docstring",
    "missing-function-docstring",
    "missing-function-docstring",
    "missing-module-docstring",
    "wrong-import-position",
    "unnecessary-ellipsis",  # Conflicts with Protocols
    "broad-except",
    "unused-argument",  # Handled by Ruff
    "redefined-builtin",  # ExceptionGroup is a builtin
]


[tool.coverage]
report.exclude_lines = [
    'pragma: no cover',
    '\.\.\.',
    'if typing.TYPE_CHECKING:',
]


[tool.check-wheel-contents]
ignore = ["W002"]  # Triggers on __init__.py's


[tool.ruff]
select = [
    "E", "F", "W", # flake8
    "B",  "B904",  # flake8-bugbear
    "I",           # isort
    "ARG",         # flake8-unused-arguments
    "C4",          # flake8-comprehensions
    "EM",          # flake8-errmsg
    "ICN",         # flake8-import-conventions
    "ISC",         # flake8-implicit-str-concat
    "PGH",         # pygrep-hooks
    "PIE",         # flake8-pie
    "PL",          # pylint
    "PT",          # flake8-pytest-style
    "PTH",         # flake8-use-pathlib
    "RET",         # flake8-return
    "RUF",         # Ruff-specific
    "SIM",         # flake8-simplify
    "T20",         # flake8-print
    "UP",          # pyupgrade
    "YTT",         # flake8-2020
]
extend-ignore = [
    "PLR0915",
    "PLR0912",
    "PLR2004",
    "PLE1205",  # Format check doesn't work with our custom logger
    "E501",  # Line too long
    "PT004",
]
target-version = "py37"
typing-modules = ["scikit_build_core._compat.typing"]
src = ["src"]
unfixable = ["T20", "F841"]
exclude = []

[tool.ruff.per-file-ignores]
"tests/**" = ["T20"]
"noxfile.py" = ["T20"]
